util de generación de codigo
