+++ CouponBook/Program.cs
@@ -115,7 +115,7 @@ builder.Services.AddScoped<IPurchaseRepository, PurchaseRepository>();
 builder.Services.AddScoped<IRedemptionRepository, RedemptionRepository>();
 builder.Services.AddScoped<IUpdateLogRepository, UpdateLogRepository>();
 builder.Services.AddScoped<GetMarketingId>();
-
+builder.Services.AddScoped<GeneradorDeCodigos>();
 
 // Agregando Scoped Services
 builder.Services.AddScoped<ICouponPermissionService, CouponPermissionService>();
